GameRule, PlayerCreation, BoardHandler updated.
GameRule has been updated in order to get numPlayer and if it is an advancedGame from id.
PlayerCreation now do not create the player anymore but collect the users' choices and check if these are valid.
BoardHandler has been updated with moveFromEntranceToIsland method.
Little updates in model to improve encapsulation