CLI updating
CLI now shows all players, error state and choices that player can do. To add these details a ErrorBean has been created and it contains the error message. UserInterface now have a list of CommandEnum, for each element of this list CLI shows the name of the command, the requirements and the number to insert in order to play that command.

Model has been updated removing deprecated constructors of SimpleGame and AdvancedGame and tests have been updated too. AdvancedGame now do not have list of AdvancedPlayer but simply add advancedPlayer in SimpleGame's Player list players.