Watcher and VirtualView.
View component of MVC has been implemented with an observer pattern. Classes who observe are Watcher and all its subclasses. Classes observed are DrawableObject's subclasses. DrawableObject has 3 methods: ToBean that create right bean; alert who notify the observers in its list that something has changed; and killAll that kill all observers (use that when this object will be not used again. Watcher has a DrawableObject that is watched, a boolean alive that show if this observer is still useful and 2 methods: update and kill. Constructor bind the object with this observer and put this observer in the right list of VirtualView (Watcher's subclasses need VirtualView as parameter of their constructors). Update, called by object when something change, get a bean from object with actual object state and save it in field bean (each Watcher subclasses have field bean of correct type). Kill set alive at false and set object's reference at null. Each DrawableObject when create itself create also a Watcher of right type and bind it with itself(the object).

Watcher name or others names have been chosen in order to not generate conflicts with previous java classes as Observer or Observable