Test & Fix
Bug in IslandGroup.build() fixed. Updating player towers was not correct. Method added towers to player when his team color was equal to winnerTeam color and, it subtracted towers to player when player's color was equal to previousTeam color. When player's color is equal to winner team, his number of towers must be decreased because we move his towers on the island. When player's color is equal to previous team, his number of color must be increased because we move his towers from island to his board.

Fixed IslandGroupWatcher notifications in SimpleGame.moveMN. Now all IslandGroupWatchers are notified.